修改参数接受类型

See merge request !562

+++ src/main/java/so/sao/shop/supplier/dao/QualificationDao.java
@@ -25,7 +25,7 @@ public interface QualificationDao {
      * @param reason
      * @param updateDate
      */
-    void updateQualificationStatus(@Param("accountId") Integer accountId, @Param("qualificationStatus")Integer qualificationStatus,@Param("reason") String reason, @Param("updateDate")Date updateDate );
+    void updateQualificationStatus(@Param("accountId") Long accountId, @Param("qualificationStatus")Integer qualificationStatus,@Param("reason") String reason, @Param("updateDate")Date updateDate );
 
     /**
      * 查询登录供应商的资质状态
@@ -66,12 +66,12 @@ public interface QualificationDao {
      * 更新资质状态消息已读状态
      * @param accountId
      */
-    void updateQualificationMessageRead(Integer accountId);
+    void updateQualificationMessageRead(Long accountId);
 
     /**
      * 判断资质消息是否已读
      * @param accountId
      * @return
      */
-    Integer findQualificationStatusIsRead(Integer accountId);
+    Integer findQualificationStatusIsRead(Long accountId);
 }
